Add files via upload
This first project is a machine learning task on the Titanic Database. Here there are some basic cleaning techniques of the NaN values and some conversion of the categorical features to numeric ones. No much analysis is provided though as the Titanic dataset was extensively researched by many individuals already and a big library exists in Kaggle.

Goals:
(1). Is it possible to create a model to predict whether a passenger surived or not the accident based on given features? 
(2). How accurately and what is the best algorithm to make this prediction?

Soure: The Titanic Dataset was imported from Kaggle library.